Crate recycle_buf

This module implements a recycler buffer system that allows for pre-allocation and memory reuse by leveraging the Rust language’s drop mechanics. The main motivation for this system is avoiding memory allocations of large and/or expensive to create objects in a realtime environment

Structs

RecycleRef

a reference to a managed T instance that allows mutation. When the reference is destroyed then the managed T instance will be recycled back into the Recycler. This reference is not thread safe but shareable thread safe references can be created by calling RecycleRef::to_shared

Recycler

A buffer that manages the recycling of dropped items that were pulled from this buffer.

RecyclerBuilder

Uses the builder pattern to pre allocate T instances and build the Recycler

SharedRecycleRef

A thread safe recycle reference that allows read access to the underlying item. Once all instances of the shared recycle reference that contain the same item are drop then the item is recycled back into the recycler buffer.